Initial CI/CD actions (#15)
Continuous Integration and Continuous Development (CI/CD) actions were
added through Github Actions, these do the following:

- The repository is able to check for builds.
- The repository notifies for deploy reviews to either branches that are
Pull Requesting to `main` or to `main` branch pushes.

All of these let us have a **Beta** environment for Quality Assurance,
before merging to the `main` branch.

The deployments done here are not meant to be used for production,
rather for testing purposes. Production code must be sent to where we
will be hosting the website.